fix(desktop): keep Advisor runtime guards within source boundary
