+++ app/renderer/sections/Advisor.tsx
@@ -12,8 +12,10 @@ import { OllamaAdvisorRuntime, probeOllama } from '../advisor/ollama'
 import { HostedAdvisorRuntime, probeHostedAdvisor } from '../advisor/hosted'
 import { periodLabel, scopeLabel } from '../advisor/evidence'
 import { advisorContextualSurfaceLabel, advisorScopeFromContextualLaunch, normalizeAdvisorContextualLaunch, type AdvisorContextualLaunchV1, type AdvisorContextualScopeMode } from '../advisor/context'
-import { advisorScopeFingerprint, type AdvisorAnswer, type AdvisorConversationTurn, type AdvisorHostedModelState, type AdvisorHostedProviderId, type AdvisorLocalRuntimeId, type AdvisorPresentationBlockV1, type AdvisorPresentationChartSeries, type AdvisorScope } from '../advisor/types'
+import { advisorScopeFingerprint, type AdvisorAnswer, type AdvisorConversationTurn, type AdvisorHostedProviderId, type AdvisorLocalRuntimeId, type AdvisorPresentationBlockV1, type AdvisorScope } from '../advisor/types'
 import { AdvisorRuntimeControls, createHostedProbeChecking, createHostedProbeFailure, presentHostedProbe, type AdvisorHostedProbePresentation, type AdvisorRuntimeChoice, type AdvisorRuntimeState } from './AdvisorRuntimeControls'
+import { AdvisorHostedOperationGuard, isSelectableHostedModel } from './advisor-hosted-operation-guard'
+import { AdvisorChartBlock } from './advisor-chart'
 type DetectedProvider = { id: string; label: string }
 type AdvisorMessage = { id: string; role: 'user' | 'assistant'; text?: string; answer?: AdvisorAnswer; scopeFingerprint: string }
 type AdvisorConversation = { id: string; title: string; messages: AdvisorMessage[] }
@@ -36,9 +38,6 @@ function isCancelled(error: unknown): boolean {
   }
   return false
 }
-function isSelectableHostedModel(model: { state: AdvisorHostedModelState }): boolean {
-  return model.state !== 'unsupported' && model.state !== 'failed-conformance'
-}
 function providerLabel(provider: string): string {
   if (provider === 'all') return 'All providers'
   return provider.split(/[-\s]+/).filter(Boolean).map(part => part.charAt(0).toUpperCase() + part.slice(1)).join(' ')
@@ -58,71 +57,11 @@ function answerForMessage(messages: AdvisorMessage[], id: string | null): Adviso
   }
   return [...messages].reverse().find(message => message.answer)?.answer ?? null
 }
-function chartValue(value: number | null, unit: string): string {
-  if (value === null || !Number.isFinite(value)) return 'Unavailable'
-  return (unit === 'USD' ? '$' + value.toFixed(2) : value.toLocaleString('en-US')) + ' ' + unit
-}
-function chartSeriesSegments(series: AdvisorPresentationChartSeries, width: number, height: number, max: number): Array<Array<[number, number]>> {
-  const segments: Array<Array<[number, number]>> = []
-  let current: Array<[number, number]> = []
-  const denominator = Math.max(1, series.points.length - 1)
-  series.points.forEach((point, index) => {
-    if (point.value === null || !Number.isFinite(point.value)) {
-      if (current.length) segments.push(current)
-      current = []
-      return
-    }
-    const x = 22 + (index / denominator) * (width - 42)
-    const y = height - 22 - (point.value / max) * (height - 42)
-    current.push([x, y])
-  })
-  if (current.length) segments.push(current)
-  return segments
-}
-
-function ChartBlock({ block }: { block: Extract<AdvisorPresentationBlockV1, { kind: 'line-chart' | 'bar-chart' }> }) {
-  const width = 620
-  const height = 190
-  const values = block.series.flatMap(series => series.points.map(point => point.value).filter((value): value is number => value !== null && Number.isFinite(value)))
-  const max = Math.max(1, ...values)
-  return (
-    <section className="advisor-presentation-block advisor-chart-block">
-      <div className="advisor-presentation-head"><h4>{block.title}</h4><span>{block.scopeLabel} · {block.periodLabel}</span></div>
-      <p className="advisor-presentation-summary">{block.summary}</p>
-      <svg className="advisor-chart" viewBox={'0 0 ' + width + ' ' + height} role="img" aria-label={block.accessibilityLabel}>
-        <line x1="22" y1={height - 22} x2={width - 20} y2={height - 22} />
-        <line x1="22" y1="18" x2="22" y2={height - 22} />
-        {block.kind === 'line-chart'
-          ? block.series.map((series, index) => <g key={series.id} className={'advisor-chart-series series-' + index}>
-              {chartSeriesSegments(series, width, height, max).map((segment, segmentIndex) => segment.length > 1 ? <polyline key={segmentIndex} points={segment.map(([x, y]) => x + ',' + y).join(' ')} /> : null)}
-              {series.points.map((point, pointIndex) => {
-                if (point.value === null || !Number.isFinite(point.value)) return null
-                const denominator = Math.max(1, series.points.length - 1)
-                const x = 22 + (pointIndex / denominator) * (width - 42)
-                const y = height - 22 - (point.value / max) * (height - 42)
-                return <circle key={pointIndex} cx={x} cy={y} r="2.6"><title>{series.label} · {point.label} · {chartValue(point.value, block.unit)}</title></circle>
-              })}
-            </g>)
-          : block.series[0]?.points.map((point, index) => {
-              if (point.value === null || !Number.isFinite(point.value)) return null
-              const slot = (width - 42) / Math.max(1, block.series[0]!.points.length)
-              const barWidth = Math.max(4, slot * .62)
-              const x = 22 + index * slot + (slot - barWidth) / 2
-              const y = height - 22 - (point.value / max) * (height - 42)
-              return <rect key={index} x={x} y={y} width={barWidth} height={Math.max(1, height - 22 - y)}><title>{point.label} · {chartValue(point.value, block.unit)}</title></rect>
-            })}
-      </svg>
-      <div className="advisor-chart-legend">{block.series.map((series, index) => <span key={series.id}><i className={'series-' + index} />{series.label}</span>)}</div>
-      <div className="advisor-chart-data">{block.series.flatMap(series => series.points.filter(point => point.value !== null).slice(-6).map(point => <span key={series.id + '-' + point.label}>{series.label} · {point.label} · {chartValue(point.value, block.unit)}</span>))}</div>
-    </section>
-  )
-}
-
 function PresentationBlocks({ blocks }: { blocks: AdvisorPresentationBlockV1[] }) {
   return <div className="advisor-presentation">{blocks.map((block, index) => {
     if (block.kind === 'text') return <p className="advisor-presentation-text" key={index}>{block.text}</p>
     if (block.kind === 'metric-cards') return <section className="advisor-presentation-block" key={index}><div className="advisor-presentation-head"><h4>{block.title}</h4><span>{block.scopeLabel} · {block.periodLabel}</span></div><div className="advisor-metric-grid">{block.cards.map(card => <div className="advisor-metric-card" key={card.label}><span>{card.label}</span><strong>{card.value}</strong><small>{card.unit}</small><p>{card.detail}</p></div>)}</div></section>
-    if (block.kind === 'line-chart' || block.kind === 'bar-chart') return <ChartBlock block={block} key={index} />
+    if (block.kind === 'line-chart' || block.kind === 'bar-chart') return <AdvisorChartBlock block={block} key={index} />
     if (block.kind === 'comparison-table') return <section className="advisor-presentation-block" key={index}><div className="advisor-presentation-head"><h4>{block.title}</h4><span>{block.scopeLabel} · {block.periodLabel}</span></div><p className="advisor-presentation-summary">{block.summary}</p><div className="advisor-table-wrap"><table><caption className="sr-only">{block.title}</caption><thead><tr>{block.table.columns.map(column => <th key={column} scope="col">{column}</th>)}</tr></thead><tbody>{block.table.rows.map((row, rowIndex) => <tr key={rowIndex}>{row.map((cell, cellIndex) => <td key={cellIndex}>{cell}</td>)}</tr>)}</tbody></table></div></section>
     if (block.kind === 'quota-card') return <section className="advisor-presentation-block" key={index}><div className="advisor-presentation-head"><h4>{block.title}</h4><span>{block.scopeLabel} · {block.periodLabel}</span></div><p className="advisor-presentation-summary">{block.summary}</p><div className="advisor-quota-grid">{block.providers.map(provider => <div className="advisor-quota-item" key={provider.provider}><strong>{provider.provider}</strong>{provider.planLabel ? <span>{provider.planLabel}</span> : null}{provider.windows.map(window => <span key={window.id}>{window.label} · {window.remainingPercent}% remaining{window.resetsAt ? ' · reset ' + window.resetsAt : ''}</span>)}{provider.creditsUSD !== null ? <span>Credits · ${provider.creditsUSD.toFixed(2)}</span> : null}</div>)}</div></section>
     if (block.kind === 'bench-summary') return <section className="advisor-presentation-block" key={index}><div className="advisor-presentation-head"><h4>{block.title}</h4><span>{block.scopeLabel} · {block.periodLabel}</span></div><p className="advisor-presentation-summary">{block.summary}</p>{block.run ? <div className="advisor-bench-summary"><strong>{block.run.model.selected}</strong><span>{block.run.aggregate.passed}/{block.run.aggregate.planned} planned tasks passed</span><span>{block.run.aggregate.scoreValue === null ? 'Score unavailable' : (block.run.aggregate.scoreValue * 100).toFixed(1) + '% score'}</span><span>Status · {block.run.status}</span></div> : <p className="advisor-muted-line">No completed controlled run is available.</p>}</section>
@@ -214,8 +153,7 @@ export function Advisor({
   const [runtimeId, setRuntimeId] = useState<AdvisorLocalRuntimeId>('ollama')
   const [runtimeChoice, setRuntimeChoice] = useState<AdvisorRuntimeChoice>('ollama')
   const [hostedProvider, setHostedProvider] = useState<'openai' | 'anthropic' | 'gemini'>('openai')
-  const hostedProviderRef = useRef(hostedProvider)
-  hostedProviderRef.current = hostedProvider
+  const hostedOperationGuardRef = useRef(new AdvisorHostedOperationGuard(hostedProvider))
   const [hostedModel, setHostedModel] = useState<string | null>(null)
   const hostedModelRef = useRef<string | null>(null)
   hostedModelRef.current = hostedModel
@@ -269,15 +207,14 @@ export function Advisor({
   }, [checkLocalRuntime])
 
   const hostedProbeController = useRef<AbortController | null>(null)
-  const hostedProbeRequestRef = useRef(0)
   const checkHostedRuntime = useCallback(async (requestedProvider: 'openai' | 'anthropic' | 'gemini' = hostedProvider, resetSelection = false) => {
-    if (hostedProviderRef.current !== requestedProvider) return
+    if (!hostedOperationGuardRef.current.isCurrentProvider(requestedProvider)) return
     hostedProbeController.current?.abort()
-    const requestId = hostedProbeRequestRef.current + 1
-    hostedProbeRequestRef.current = requestId
+    const requestId = hostedOperationGuardRef.current.startProbe(requestedProvider)
+    if (requestId === null) return
     const controller = new AbortController()
     hostedProbeController.current = controller
-    const isCurrentRequest = () => !controller.signal.aborted && hostedProviderRef.current === requestedProvider && hostedProbeRequestRef.current === requestId
+    const isCurrentRequest = () => !controller.signal.aborted && hostedOperationGuardRef.current.isCurrentProbe(requestedProvider, requestId)
     setHostedProbe(current => createHostedProbeChecking(requestedProvider, current))
     try {
       const result = await probeHostedAdvisor(requestedProvider, controller.signal)
@@ -313,7 +250,6 @@ export function Advisor({
   const [loadingQuestion, setLoadingQuestion] = useState<string | null>(null)
   const [credentialEntry, setCredentialEntry] = useState('')
   const [credentialSaving, setCredentialSaving] = useState(false)
-  const credentialOperationRef = useRef(0)
   const hostedConfigRef = useRef({ runtimeChoice, hostedRuntime, hostedConsent })
   hostedConfigRef.current = { runtimeChoice, hostedRuntime, hostedConsent }
   useEffect(() => { setCredentialEntry('') }, [hostedProvider])
@@ -451,9 +387,7 @@ export function Advisor({
     setHostedConsent(consent)
   }
   const updateHostedProvider = (next: AdvisorHostedProviderId) => {
-    hostedProviderRef.current = next
-    hostedProbeRequestRef.current += 1
-    credentialOperationRef.current += 1
+    hostedOperationGuardRef.current.setProvider(next)
     hostedProbeController.current?.abort()
     setHostedProvider(next)
     setHostedModel(null)
@@ -479,18 +413,18 @@ export function Advisor({
   const saveHostedCredential = async () => {
     if (!credentialEntry.trim() || credentialSaving) return
     const requestedProvider = hostedProvider
-    const operationId = credentialOperationRef.current + 1
-    credentialOperationRef.current = operationId
+    const operationId = hostedOperationGuardRef.current.startCredential(requestedProvider)
+    if (operationId === null) return
     setCredentialSaving(true)
     try {
       const status = await metrora.advisorCredentialSet(requestedProvider, credentialEntry)
-      if (credentialOperationRef.current !== operationId || hostedProviderRef.current !== requestedProvider) return
+      if (!hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) return
       setNotice(status.state === 'ready' ? 'Provider credential saved in protected local storage.' : 'Provider credential was not saved: ' + status.state + '.')
       await checkHostedRuntime(requestedProvider)
     } catch {
-      if (credentialOperationRef.current === operationId && hostedProviderRef.current === requestedProvider) setNotice('Provider credential could not be saved. Enter it again.')
+      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) setNotice('Provider credential could not be saved. Enter it again.')
     } finally {
-      if (credentialOperationRef.current === operationId) {
+      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) {
         setCredentialEntry('')
         setCredentialSaving(false)
       }
@@ -498,16 +432,16 @@ export function Advisor({
   }
   const clearHostedCredential = async () => {
     const requestedProvider = hostedProvider
-    const operationId = credentialOperationRef.current + 1
-    credentialOperationRef.current = operationId
+    const operationId = hostedOperationGuardRef.current.startCredential(requestedProvider)
+    if (operationId === null) return
     try {
       await metrora.advisorCredentialClear(requestedProvider)
-      if (credentialOperationRef.current !== operationId || hostedProviderRef.current !== requestedProvider) return
+      if (!hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) return
       setHostedConsent(false)
       setNotice('Provider credential removed from this device.')
       await checkHostedRuntime(requestedProvider)
     } catch {
-      if (credentialOperationRef.current === operationId && hostedProviderRef.current === requestedProvider) setNotice('Provider credential could not be removed.')
+      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) setNotice('Provider credential could not be removed.')
     }
   }
   const normalizedHistoryQuery = historyQuery.trim().toLowerCase()
